fraud-detection/requirements.txt
# Core dependencies
pandas>=2.3.3
numpy>=2.3.5
scikit-learn>=1.8.0
xgboost>=3.1.2
lightgbm>=4.6.0
shap>=0.50.0
imbalanced-learn>=0.14.1
matplotlib>=3.10.8
seaborn>=0.13.2

# Development tools
jupyter>=1.1.1
black>=25.12.0
flake8>=7.3.0
pytest>=9.0.2

# Utilities
python-json-logger>=4.0.0
PyYAML>=6.0.3
tqdm>=4.67.1
